+++ docs/pages/api-docs/sub-menu.json
@@ -3,9 +3,8 @@
   "name": "SubMenu",
   "styles": { "classes": [], "globalClasses": {}, "name": null },
   "spread": true,
-  "filename": "/packages/material-ui/src/SubMenu/SubMenu.js",
+  "filename": "/packages/mui-material/src/SubMenu/SubMenu.js",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/components/menus/\">Menus</a></li></ul>",
-  "styledComponent": true,
   "cssComponent": false
 }
