--- packages/material-ui/src/Menu/Menu.d.ts
@@ -1,95 +1,0 @@
-import * as React from 'react';
-import { SxProps } from '@material-ui/system';
-import { InternalStandardProps as StandardProps } from '..';
-import { PopoverProps } from '../Popover';
-import { MenuListProps } from '../MenuList';
-import { Theme } from '../styles';
-import { TransitionProps } from '../transitions/transition';
-import { MenuClasses } from './menuClasses';
-
-export interface MenuProps extends StandardProps<PopoverProps> {
-  /**
-   * An HTML element, or a function that returns one.
-   * It's used to set the position of the menu.
-   */
-  anchorEl?: PopoverProps['anchorEl'];
-  /**
-   * If `true` (Default) will focus the `[role="menu"]` if no focusable child is found. Disabled
-   * children are not focusable. If you set this prop to `false` focus will be placed
-   * on the parent modal container. This has severe accessibility implications
-   * and should only be considered if you manage focus otherwise.
-   * @default true
-   */
-  autoFocus?: boolean;
-  /**
-   * Menu contents, normally `MenuItem`s.
-   */
-  children?: React.ReactNode;
-  /**
-   * Override or extend the styles applied to the component.
-   */
-  classes?: Partial<MenuClasses>;
-  /**
-   * When opening the menu will not focus the active item but the `[role="menu"]`
-   * unless `autoFocus` is also set to `false`. Not using the default means not
-   * following WAI-ARIA authoring practices. Please be considerate about possible
-   * accessibility implications.
-   * @default false
-   */
-  disableAutoFocusItem?: boolean;
-  /**
-   * Props applied to the [`MenuList`](/api/menu-list/) element.
-   * @default {}
-   */
-  MenuListProps?: Partial<MenuListProps>;
-  /**
-   * Callback fired when the component requests to be closed.
-   *
-   * @param {object} event The event source of the callback.
-   * @param {string} reason Can be: `"escapeKeyDown"`, `"backdropClick"`, `"tabKeyDown"`.
-   */
-  onClose?: PopoverProps['onClose'];
-  /**
-   * If `true`, the component is shown.
-   */
-  open: boolean;
-  /**
-   * `classes` prop applied to the [`Popover`](/api/popover/) element.
-   */
-  PopoverClasses?: PopoverProps['classes'];
-  setParentOpenSubMenuIndex?: (index: number) => void;
-  /**
-   * The system prop that allows defining system overrides as well as additional CSS styles.
-   */
-  sx?: SxProps<Theme>;
-  /**
-   * The length of the transition in `ms`, or 'auto'
-   * @default 'auto'
-   */
-  transitionDuration?: TransitionProps['timeout'] | 'auto';
-  /**
-   * Props applied to the transition element.
-   * By default, the element is based on this [`Transition`](http://reactcommunity.org/react-transition-group/transition) component.
-   * @default {}
-   */
-  TransitionProps?: TransitionProps;
-  /**
-   * The variant to use. Use `menu` to prevent selected items from impacting the initial focus.
-   * @default 'selectedMenu'
-   */
-  variant?: 'menu' | 'selectedMenu';
-}
-
-/**
- *
- * Demos:
- *
- * - [App Bar](https://material-ui.com/components/app-bar/)
- * - [Menus](https://material-ui.com/components/menus/)
- *
- * API:
- *
- * - [Menu API](https://material-ui.com/api/menu/)
- * - inherits [Popover API](https://material-ui.com/api/popover/)
- */
-export default function Menu(props: MenuProps): JSX.Element;
