--- packages/material-ui/src/MenuItem/menuItemClasses.ts
@@ -1,39 +1,0 @@
-import { generateUtilityClass, generateUtilityClasses } from '@material-ui/unstyled';
-
-export interface MenuItemClasses {
-  /** Styles applied to the root element. */
-  root: string;
-  /** State class applied to the root element if keyboard focused. */
-  focusVisible: string;
-  /** Styles applied to the root element if dense. */
-  dense: string;
-  /** State class applied to the root element if `disabled={true}`. */
-  disabled: string;
-  /** Styles applied to the root element if `divider={true}`. */
-  divider: string;
-  /** Styles applied to the inner `component` element unless `disableGutters={true}`. */
-  gutters: string;
-  /** State class applied to the root element if `selected={true}`. */
-  selected: string;
-  /** State class applied to the root element if its submenu is open. */
-  openSubMenuParent: string;
-}
-
-export type MenuItemClassKey = keyof MenuItemClasses;
-
-export function getMenuItemUtilityClass(slot: string): string {
-  return generateUtilityClass('MuiMenuItem', slot);
-}
-
-const menuItemClasses: MenuItemClasses = generateUtilityClasses('MuiMenuItem', [
-  'root',
-  'focusVisible',
-  'dense',
-  'disabled',
-  'divider',
-  'gutters',
-  'selected',
-  'openSubMenuParent',
-]);
-
-export default menuItemClasses;
