+++ packages/mui-material/src/Menu/Menu.test.js
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { spy, useFakeTimers } from 'sinon';
 import { expect } from 'chai';
 import {
+  act,
   createClientRender,
   describeConformance,
   screen,
@@ -10,6 +11,9 @@ import {
 } from 'test/utils';
 import Menu, { menuClasses as classes } from '@mui/material/Menu';
 import Popover from '@mui/material/Popover';
+import Button from '../Button';
+import SubMenu from '../SubMenu';
+import MenuItem from '../MenuItem';
 
 describe('<Menu />', () => {
   /**
@@ -156,7 +160,7 @@ describe('<Menu />', () => {
   });
 
   it('should open during the initial mount', () => {
-    function MenuItem(props) {
+    function MenuItemMock(props) {
       const { autoFocus, children } = props;
       return (
         <div role="menuitem" tabIndex={-1} data-autofocus={autoFocus}>
@@ -166,7 +170,7 @@ describe('<Menu />', () => {
     }
     render(
       <Menu anchorEl={document.createElement('div')} open>
-        <MenuItem>one</MenuItem>
+        <MenuItemMock>one</MenuItemMock>
       </Menu>,
     );
 
@@ -211,7 +215,7 @@ describe('<Menu />', () => {
   });
 
   it('should call onClose on tab', () => {
-    function MenuItem(props) {
+    function MenuItemMock(props) {
       const { autoFocus, children } = props;
 
       const ref = React.useRef(null);
@@ -230,7 +234,7 @@ describe('<Menu />', () => {
     const onCloseSpy = spy();
     render(
       <Menu anchorEl={document.createElement('div')} open onClose={onCloseSpy}>
-        <MenuItem>hello</MenuItem>
+        <MenuItemMock>hello</MenuItemMock>
       </Menu>,
     );
 
@@ -268,6 +272,752 @@ describe('<Menu />', () => {
         !strictModeDoubleLoggingSupressed &&
           "MUI: The Menu component doesn't accept a Fragment as a child.",
       ]);
+    });
+  });
+
+  describe('cascading menu', () => {
+    it('renders a subMenu', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseMove(getByRole('menuitem', { name: 'Test' }));
+      });
+      clock.restore();
+      expect(getByRole('menuitem', { name: expected })).to.not.equal(null);
+    });
+
+    it('renders a nested subMenu', () => {
+      const expected = 'NestedSubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu transitionDuration={0}>
+                    <MenuItem
+                      subMenu={
+                        <SubMenu>
+                          <MenuItem>{expected}</MenuItem>
+                        </SubMenu>
+                      }
+                    >
+                      Test2
+                    </MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseMove(getByRole('menuitem', { name: 'Test' }));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseMove(getByRole('menuitem', { name: 'Test2' }));
+      });
+
+      clock.restore();
+      expect(getByRole('menuitem', { name: expected })).to.not.equal(null);
+    });
+
+    it('collapses the subMenu when active parent item is changed', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+              <MenuItem>Other</MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole, queryByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseMove(getByRole('menuitem', { name: 'Test' }));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseMove(getByRole('menuitem', { name: 'Other', hidden: true }));
+      });
+
+      clock.restore();
+      expect(queryByRole('menuitem', { name: expected })).to.equal(null);
+    });
+
+    it('keeps subMenus open when mousing outside of menus', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole, queryByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseMove(getByRole('menuitem', { name: 'Test' }));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseOut(getByRole('menuitem', { name: 'Test', hidden: true }));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.mouseEnter(getByRole('button', { hidden: true }));
+      });
+
+      clock.restore();
+      expect(queryByRole('menuitem', { name: expected })).to.not.equal(null);
+    });
+
+    it('opens a subMenu on right arrow keydown', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole, queryByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'Test' }), { key: 'ArrowRight' });
+      });
+
+      clock.restore();
+      expect(queryByRole('menuitem', { name: expected })).to.not.equal(null);
+    });
+
+    it('closes a subMenu on left arrow keydown', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole, queryByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'Test' }), { key: 'ArrowRight' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: expected }), { key: 'ArrowLeft' });
+      });
+
+      clock.restore();
+      expect(queryByRole('menuitem', { name: expected })).to.equal(null);
+    });
+
+    it('closes all menus on tab keydown', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        const handleClose = () => {
+          setAnchorEl(null);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu
+              anchorEl={anchorEl}
+              onClose={handleClose}
+              open={Boolean(anchorEl)}
+              transitionDuration={0}
+            >
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole, queryByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'Test' }), { key: 'ArrowRight' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: expected }), { key: 'Tab' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+      clock.restore();
+
+      expect(queryByRole('menuitem', { name: expected })).to.equal(null);
+      expect(queryByRole('menuitem', { name: 'Test' })).to.equal(null);
+    });
+
+    it('closes all menus on escape keydown', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        const handleClose = () => {
+          setAnchorEl(null);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu
+              anchorEl={anchorEl}
+              onClose={handleClose}
+              open={Boolean(anchorEl)}
+              transitionDuration={0}
+            >
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole, queryByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'Test' }), { key: 'ArrowRight' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: expected }), { key: 'Escape' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      clock.restore();
+
+      expect(queryByRole('menuitem', { name: expected })).to.equal(null);
+      expect(queryByRole('menuitem', { name: 'Test' })).to.equal(null);
+    });
+
+    it('changes subMenu item focus with down arrow', () => {
+      const expected = 'Second';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>First</MenuItem>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'Test' }), { key: 'ArrowRight' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'First' }), { key: 'ArrowDown' });
+      });
+
+      clock.restore();
+      expect(getByRole('menuitem', { name: expected })).to.equal(document.activeElement); // is focused
+      expect(Array.from(getByRole('menuitem', { name: expected }).classList)).to.include(
+        'Mui-focusVisible',
+      ); // looks focused
+    });
+
+    it('changes subMenu item focus with up arrow', () => {
+      const expected = 'Second';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>First</MenuItem>
+                    <MenuItem>{expected}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'Test' }), { key: 'ArrowRight' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'First' }), { key: 'ArrowUp' });
+      });
+
+      clock.restore();
+      expect(getByRole('menuitem', { name: expected })).to.equal(document.activeElement); // is focused
+      expect(Array.from(getByRole('menuitem', { name: expected }).classList)).to.include(
+        'Mui-focusVisible',
+      ); // looks focused
+    });
+
+    it('focuses first item when it opens a subMenu', () => {
+      const expected = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{expected}</MenuItem>
+                    <MenuItem>Other</MenuItem>
+                  </SubMenu>
+                }
+              >
+                Test
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: 'Test' }), { key: 'ArrowRight' });
+      });
+
+      clock.restore();
+      expect(getByRole('menuitem', { name: expected })).to.equal(document.activeElement); // is focused
+      expect(Array.from(getByRole('menuitem', { name: expected }).classList)).to.include(
+        'Mui-focusVisible',
+      ); // looks focused
+    });
+
+    it('changes focus with right and left arrow keys', () => {
+      const firstFocus = 'MenuItem';
+      const secondFocus = 'SubMenuItem';
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>{secondFocus}</MenuItem>
+                  </SubMenu>
+                }
+              >
+                {firstFocus}
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { getByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(getByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      // ensure focus is on first item in root menu
+      expect(getByRole('menuitem', { name: firstFocus })).to.equal(document.activeElement); // is focused
+      expect(Array.from(getByRole('menuitem', { name: firstFocus }).classList)).to.include(
+        'Mui-focusVisible',
+      ); // looks focused
+
+      // arrow right
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: firstFocus }), { key: 'ArrowRight' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      // ensure focus moved to first item in submenu
+      expect(getByRole('menuitem', { name: secondFocus })).to.equal(document.activeElement); // is focused
+      expect(Array.from(getByRole('menuitem', { name: secondFocus }).classList)).to.include(
+        'Mui-focusVisible',
+      ); // looks focused
+
+      // arrow back left
+      act(() => {
+        fireEvent.keyDown(getByRole('menuitem', { name: secondFocus }), { key: 'ArrowLeft' });
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      // ensure focus moved back to first item in root menu
+      expect(getByRole('menuitem', { name: firstFocus })).to.equal(document.activeElement); // is focused
+      expect(Array.from(getByRole('menuitem', { name: firstFocus }).classList)).to.include(
+        'Mui-focusVisible',
+      ); // looks focused
+    });
+
+    it('keeps parent items of open sub menus highlighted', () => {
+      const CascadingMenu = () => {
+        const [anchorEl, setAnchorEl] = React.useState(null);
+
+        const handleButtonClick = (event) => {
+          setAnchorEl(event.currentTarget);
+        };
+
+        return (
+          <React.Fragment>
+            <Button onClick={handleButtonClick} />
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} transitionDuration={0}>
+              <MenuItem
+                subMenu={
+                  <SubMenu>
+                    <MenuItem>SubMenuItem</MenuItem>
+                  </SubMenu>
+                }
+              >
+                MenuItem
+              </MenuItem>
+            </Menu>
+          </React.Fragment>
+        );
+      };
+
+      const { queryByRole } = render(<CascadingMenu />);
+
+      act(() => {
+        fireEvent.click(queryByRole('button'));
+      });
+
+      act(() => {
+        clock.tick(0);
+      });
+
+      act(() => {
+        fireEvent.keyDown(queryByRole('menuitem', { name: 'MenuItem' }), { key: 'ArrowRight' });
+      });
+
+      clock.restore();
+
+      const classList = Array.from(
+        queryByRole('menuitem', { name: 'MenuItem', hidden: true }).classList,
+      );
+      const findClassLike = (matchString, list) => {
+        const matches = list.filter((c) => c.indexOf(matchString) > -1);
+        return matches.length > 0;
+      };
+
+      expect(findClassLike('openSubMenuParent', classList)).to.equal(true);
     });
   });
 });
